20250617-183757-995774300

+++ cmd/copy/copy.go
@@ -112,11 +112,13 @@ for more info.
 				// 定义回调函数，在文件复制完成后删除源文件
 				callback := func(obj fs.Object) error {
 					if deleteAfterCopy {
-						// 只要复制成功就删除源文件，不检查大小
-						fs.Debugf(obj, "Deleting source file after successful copy")
-						time.Sleep(5 * time.Second)
-						return operations.DeleteFile(context.Background(), obj)
-					}
+						// 检查目标文件是否存在且大小相同
+						dstObj, err := fdst.NewObject(context.Background(), obj.Remote())
+						if err == nil && dstObj != nil && obj.Size() == dstObj.Size() {
+							fs.Debugf(obj, "Deleting source file as it has same size as destination")
+							time.Sleep(3 * time.Second)
+							return operations.DeleteFile(context.Background(), obj)
+						}
 					return nil
 				}
 				// 修改CopyDir调用，添加noCheckDest参数
@@ -133,7 +135,7 @@ for more info.
 				// 如果目标文件存在且大小相同，则删除源文件
 				if deleteAfterCopy && obj.Size() == dstObj.Size() {
 					fs.Debugf(obj, "Deleting source file as it has same size as destination")
-					time.Sleep(2 * time.Second)
+					time.Sleep(3* time.Second)
 					return operations.DeleteFile(context.Background(), obj)
 				}
 				fs.Debugf(obj, "Skipping as unchanged")
@@ -141,7 +143,7 @@ for more info.
 			}
 			err = operations.CopyFile(context.Background(), fdst, fsrc, srcFileName, srcFileName)
 			if err == nil && deleteAfterCopy {
-				time.Sleep(5 * time.Second)
+				time.Sleep(3 * time.Second)
 				return operations.DeleteFile(context.Background(), obj)
 			}
 			return err
